pyproject.toml
[tool.poetry]
name = "fastapi-todo"
version = "0.1.0"
description = ""
authors = ["Diego Ubirajara <diego.ubirajara@mail.com>"]
readme = "README.md"
packages = [{include = "fastapi_todo"}]

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
pydantic = {extras = ["email"], version = "^2.7.4"}
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"


[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]


[tool.ruff.lint]
preview = true
select = ["E", "F", "I", "W", "PL", "PT"]


[tool.ruff.format]
preview = true
quote-style = "single"


[tool.taskipy.tasks]
run = "fastapi dev fastapi_todo/app.py"
lint = "ruff check . && ruff check . --diff"
format = "ruff check . --fix && ruff format ."
pre_test = "task lint"
test = "pytest -s -x --cov=fastapi_todo -vv"
post_test = "coverage html"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
